moved types to the reconciler
On-behalf-of: @SAP a.shcherbatiuk@sap.com
